# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RecallAI
  version: 0.1.0
paths:
...
  /api/v1/vault/save:
    post:
      tags:
      - vault
      summary: Save Url
      description: 'Save a URL. Returns immediately; AI runs async in the worker.


        Answers 200 instead of 201 when the link is already in the vault, so the client can

        say "already saved" rather than implying a second copy was created.'
      operationId: save_url_api_v1_vault_save_post
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveUrlRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/note:
    post:
      tags:
      - vault
      summary: Save Note
      operationId: save_note_api_v1_vault_note_post
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/upload:
    post:
      tags:
      - vault
      summary: Upload Document
      description: 'Store an uploaded document in the bucket; index its text when it has any.


        Read with an explicit cap rather than trusting `file.size`: the client controls the

        Content-Length header, so a small declared size can still stream a large body. The

        file type is decided from the bytes inside `services/documents.py`, never from the

        filename or the browser''s Content-Type.'
      operationId: upload_document_api_v1_vault_upload_post
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_v1_vault_upload_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/voice:
    post:
      tags:
      - vault
      summary: Save Voice Note
      description: 'Transcribe a spoken note and save what was said.


        The transcript is the memory: it is summarised, tagged and embedded like any other

        item, so a voice note is searchable by its words rather than by its filename. The

        recording is filed alongside it when a bucket is configured, and its absence never

        costs the user the transcript.


        Read with an explicit cap rather than trusting `audio.size` -- the client writes

        Content-Length, so a small declared size can still stream a large body. The container

        is decided from the bytes in `services/transcription.py`; the filename a browser

        invents for a `MediaRecorder` blob is never consulted.


        `language` (ISO-639-1) pins the transcription instead of letting the model detect it.

        Auto-detection is the default and is right for a multilingual vault, but it *is* a

        guess -- a short Bengali clip has been observed coming back as fluent Traditional

        Chinese -- so the recorder offers the choice and this removes the guess entirely.

        Unknown codes fall back to auto-detect rather than being refused: the value is

        forwarded to a provider and rendered on a page, so it is re-derived from a closed

        allowlist, and a client bug must not cost the user their words.


        `duration` is the recorder''s own measurement, kept because only the whisper-* models

        report one and the player needs a length regardless.


        `peaks` is the amplitude shape the recorder measured while recording, used to draw the

        waveform on the memory page. It is client-written data bound for a JSONB column, so

        `parse_waveform` re-derives it rather than trusting it and answers None for anything

        malformed -- the picture is decoration beside a transcript and must never cost the

        user the words.'
      operationId: save_voice_note_api_v1_vault_voice_post
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_save_voice_note_api_v1_vault_voice_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/uploads/limits:
    get:
      tags:
      - vault
      summary: Upload Limits
      description: 'What the client may send, so the picker can filter before a 25MB round-trip.


        This is deployment configuration -- it changes when the server restarts, never

        between two requests -- but the picker asks for it on every mount, and against a

        database in another region that was a ~600ms wait before the user could choose a

        file. `private` keeps it out of shared caches (it is behind a session cookie and

        describes that user''s own limits); the max-age is short enough that a config change

        is picked up within a minute of a deploy.'
      operationId: upload_limits_api_v1_vault_uploads_limits_get
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Upload Limits Api V1 Vault Uploads Limits Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault:
    get:
      tags:
      - vault
      summary: List Vault
      operationId: list_vault_api_v1_vault_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/{item_id}:
    get:
      tags:
      - vault
      summary: Get Item
      operationId: get_item_api_v1_vault__item_id__get
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItemDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - vault
      summary: Delete Item
      operationId: delete_item_api_v1_vault__item_id__delete
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/{item_id}/content:
    patch:
      tags:
      - vault
      summary: Update Item Content
      description: 'Replace this item''s body with what the user wrote in the editor.


        The request carries blocks and nothing else -- the plain text, the stored document

        and the surviving highlights are all derived server-side, so the browser cannot

        write a `content` that disagrees with the document it sent, nor touch any other

        column by adding it to the body.


        404 covers "no such item" and "not yours" alike: the repository scopes on `user_id`,

        so a guessed id from another account is indistinguishable from a missing one.'
      operationId: update_item_content_api_v1_vault__item_id__content_patch
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItemDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/{item_id}/reprocess:
    post:
      tags:
      - vault
      summary: Reprocess Item
      description: 'Put a failed or skipped item back on the queue.


        The pipeline can fail for reasons that have nothing to do with the item -- a provider

        timeout, a rate limit, a worker that was killed mid-job -- and until now the only

        thing a person could do about it was delete the memory and save it again, losing the

        id, the link and anything they had edited. This is the same work, re-driven.


        404 covers "no such item" and "not yours" alike: the repository scopes on `user_id`,

        so a guessed id from another account is indistinguishable from a missing one.


        409 is the honest answer for an item that is already queued or already finished --

        the client hides the button in both states, so reaching here means a double click.

        429 is the cooldown. Neither is an error worth a stack trace.


        A **voice note** is the one kind that may be re-driven after it *succeeded*, because

        its transcript is the one output that can be confidently and fluently wrong, and its

        audio is still in the bucket to redo it from. `language` (ISO-639-1) pins that re-run,

        which is the point: repeating a failed auto-detection unchanged is the same coin flip.'
      operationId: reprocess_item_api_v1_vault__item_id__reprocess_post
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/ReprocessRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vault/{item_id}/file:
    get:
      tags:
      - vault
      summary: Get Item File
      description: 'Mint a short-lived download URL for this item''s stored file.


        404 covers "no such item", "not yours" and "has no file" alike -- the repository

        scopes on `user_id`, so another account''s id is indistinguishable from a missing one

        and the endpoint cannot be used to probe which ids exist.'
      operationId: get_item_file_api_v1_vault__item_id__file_get
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileLinkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_document_api_v1_vault_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_document_api_v1_vault_upload_post
...
    ContentType:
      type: string
      enum:
      - youtube
      - article
      - pdf
      - document
      - note
      - instagram
      - facebook
      - tiktok
      - linkedin
      - voice
      - image
      title: ContentType
...
    CreateNoteRequest:
      properties:
        title:
          type: string
          maxLength: 512
          minLength: 1
          title: Title
        content:
          type: string
          minLength: 1
          title: Content
      type: object
      required:
      - title
      - content
      title: CreateNoteRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SaveUrlRequest:
      properties:
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      required:
      - url
      title: SaveUrlRequest
...
    VaultItemRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        type:
          $ref: '#/components/schemas/ContentType'
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        thumbnail_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Url
        ai_tags:
          items:
            type: string
          type: array
          title: Ai Tags
        ai_category:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Category
        ai_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Label
        processing_status:
          $ref: '#/components/schemas/ProcessingStatus'
        processing_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Processing Error
        created_at:
          type: string
          format: date-time
          title: Created At
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
        file_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: File Size
        mime_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Mime Type
      type: object
      required:
      - id
      - type
      - source_url
      - title
      - summary
      - thumbnail_url
      - ai_tags
      - ai_category
      - processing_status
      - created_at
      title: VaultItemRead
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: recall_session
      description: Signed JWT session cookie set by /api/v1/auth/google/callback.